Extract url-arbiter client into a separate package
