+++ urlarbiter/url_arbiter.go
@@ -1,4 +1,4 @@
-package main
+package urlarbiter
 
 import (
 	"bytes"
